pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "scengen"
version = "0.1.0"
description = "Scenario generation and latent space analysis app."
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
  "flask",
  "dash",
  "flask_sqlalchemy",
  "torch",
  "scikit-learn",
  "pandas",
  "plotly",
  "minisom",
  "redis",
  "requests",
  "python-dotenv",
  "psycopg2-binary"
]

[tool.setuptools.packages.find]
include = [
  "Pytorch", "Pytorch.*",
  "Backend", "Backend.*",
  "Shared", "Shared.*",
  "scripts", "scripts.*",
]

[project.scripts]
fetch-fx      = "scripts.fetch_fx:main"
start-backend = "Backend.app.run:main"
